fixed warnings

+++ Classy/Additions/UIColor+CASAdditions.m
@@ -14,7 +14,7 @@
 
 @implementation NSString (CASPrivateAdditions)
 - (NSUInteger)cas_hexValue {
-    NSUInteger result = 0;
+    unsigned long result = 0;
     sscanf([self UTF8String], "%lx", &result);
     return result;
 }
